feat: add tbigint type + th3index temporal H3 index type (Phase 6 compliant)

+++ meos/src/temporal/span_meos.c
@@ -326,7 +326,7 @@ int_to_span(int i)
  * @csqlfn #Value_to_span()
  */
 Span *
-bigint_to_span(int i)
+bigint_to_span(int64 i)
 {
   Span *result = palloc(sizeof(Span));
   /* Account for canonicalized spans */
